+++ src/main/java/com/axelor/studio/service/builder/StudioActionViewService.java
@@ -27,7 +27,7 @@ import com.axelor.studio.service.StudioMetaService;
 import com.google.common.base.Strings;
 import com.google.inject.Inject;
 import java.util.List;
-import org.apache.commons.lang3.StringEscapeUtils;
+import org.apache.commons.text.StringEscapeUtils;
 
 public class StudioActionViewService {
 
@@ -74,7 +74,7 @@ public class StudioActionViewService {
 
     for (StudioActionLine param : params) {
       xml.append("\n" + INDENT + "<view-param name=\"" + param.getName() + "\" ");
-      xml.append("value=\"" + StringEscapeUtils.escapeXml(param.getValue()) + "\" />");
+      xml.append("value=\"" + StringEscapeUtils.escapeXml11(param.getValue()) + "\" />");
     }
   }
 
@@ -92,9 +92,9 @@ public class StudioActionViewService {
             || contextValue.startsWith("action:")
             || contextValue.startsWith("select:")
             || contextValue.startsWith("select[]:")) {
-          xml.append("expr=\"" + StringEscapeUtils.escapeXml(contextValue) + "\" />");
+          xml.append("expr=\"" + StringEscapeUtils.escapeXml11(contextValue) + "\" />");
         } else {
-          xml.append("expr=\"eval:" + StringEscapeUtils.escapeXml(contextValue) + "\" />");
+          xml.append("expr=\"eval:" + StringEscapeUtils.escapeXml11(contextValue) + "\" />");
         }
       }
     }
@@ -117,7 +117,7 @@ public class StudioActionViewService {
     }
 
     if (domain != null) {
-      xml.append("\n" + INDENT + "<domain>" + StringEscapeUtils.escapeXml(domain) + "</domain>");
+      xml.append("\n" + INDENT + "<domain>" + StringEscapeUtils.escapeXml11(domain) + "</domain>");
     }
   }